Add initial support for ndarrays
PyArrow doesn't natively support multi-dimensional arrays. Here we
modify the `load_array_tsv` loader to return a struct that contains the
flattened array `data` and the original array `shape`. This way, the
original array can be reconstructed.

One complication is that PyArrow can't cast to these complex struct
types. So we do away with the unnecessary type casting within the
handler altogether. (The schema is still enforced when the
`IncrementalTable` is converted to a `pa.Table`.)